+++ src/init.cpp
@@ -1070,7 +1070,7 @@ bool AppInitParameterInteraction()
     if (mapArgs.count("-minrelaytxfee"))
     {
         CAmount n = 0;
-        if (!ParseMoney(mapArgs["-minrelaytxfee"], n))
+        if (!ParseMoney(mapArgs["-minrelaytxfee"], n) || 0 == n)
             return InitError(AmountErrMsg("minrelaytxfee", mapArgs["-minrelaytxfee"]));
         // High fee check is done afterward in CWallet::ParameterInteraction()
         ::minRelayTxFee = CFeeRate(n);